Move `ddcprimitives` content to `pallets` package
